deleted old vision code and cleaned up for review
